refactor: reconstruct modal directory structure
- move ModalPortal and ModalWindow to @common
- move provider to hook
- combine ModalWindow and ModalPortal in ModalProvider

Refs: https://github.com/woowacourse/service-apply/pull/649#discussion_r990600732
