+++ frontend/src/components/ApplicationPreviewModal/ApplicationPreviewModal.module.css
@@ -2,6 +2,7 @@
   display: flex;
   flex-direction: column;
   gap: 1.5rem;
+  overflow: hidden;
 }
 
 .title {
